refactor: Map api 코드를 수정한다. - 지도의 크기를 변경
